Move Functional Component from [id].tsx
This commit simplifies [id].tsx by removing how the
blog post is rendered from the Blog Post itself.

This is done to make the [id].tsx simpler to
understand as well as to separate responsibilities.
The [id].tsx should only be focused on the logic
but the components folder should be focused on the
actual looks